pyproject.toml
[project]
name = "spectrseqtools"
version = "0.1.0"
description = ""
authors = [
    { name = "Laura Christine Kuehle", email = "laura.kuehle@mpi-dortmund.mpg.de" },
    { name = "Juan Antonio Magalang", email = "juan.magalang@mpi-dortmund.mpg.de" },
    { name = "Moshir Harsh", email = "btemoshir@gmail.com" },
    { name = "Johannes Koester", email = "johannes.koester@uni-due.de" }
]
readme = "README.md"
license = "GPL-3.0-only"
requires-python=">=3.12,<3.14"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]

dependencies = [
    "polars >=1.9.0,<2.0",
    "pulp >=2.3.0,<3.0",
    "typed-argument-parser >=1.10.1,<2.0",
    "altair >=5.4.1,<6.0",
    "pyyaml >=6.0.2,<7.0",
    "loguru >=0.7.2,<0.8",
    "numpy >=2.2.3,<3.0",
    "platformdirs >=4.3.8,<5.0",
    "clr-loader >=0.2.7.post0,<0.3",
    "ms-deisotope >=0.0.60,<0.1",
    "pythonnet >=3.0.5,<4.0",
    "tqdm >=4.67.1,<5.0",
    "scikit-learn >=1.7.2,<2.0",
    "dbscan1d >=0.2.3,<0.3",
    "matplotlib >=3.10.6,<4.0",
    "cython >=3.0.3,<4.0"
]

[project.scripts]
spectrseqtools = "spectrseqtools.cli:main"

[project.urls]
Homepage = "https://github.com/spectrseq/spectrseqtools"

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64", "linux-aarch64", "osx-64"]

[tool.pixi.dependencies]
mono = ">=6.12.0.199,<7"

[tool.pixi.feature.dev.dependencies]
ruff = "*"
pytest = "*"

[tool.pixi.feature.dev.tasks]
format = "ruff format"
lint = "ruff check"
test = "pytest"

[tool.pixi.environments]
dev = {features = ["dev"]}

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

